requirements.txt
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.5
django-debug-toolbar==1.11
django-extensions==2.1.4
django-jinja==2.4.1
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mysqlclient==1.4.2.post1
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.0
openpyxl==2.5.14
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.2
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psycopg2==2.7.7
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
pywinpty==0.5.5
pyzmq==17.1.2
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.2.4
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
